backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.36
asyncpg==0.29.0
psycopg[binary]==3.2.1
python-dotenv==1.0.1
Pillow==10.4.0
apscheduler==3.10.4
pytz==2024.1
bcrypt==4.1.2
